+++ ExhibitGrid/ExhibitGrid/ViewModel/RowVm.cs
@@ -10,7 +10,7 @@ namespace ExhibitGrid.ViewModel
         public string GridCode { get; set; }
         public string RowCode { get; set; }
         public string ParentRowCode { get; set; }
-        public int DisplayOrder { get; set; }
+        public decimal DisplayOrder { get; set; }
         public bool IsHidden { get; set; }
         public bool IsCollapsed { get; set; }
         public string Class { get; set; }
